fix(schema): hide traceback in CLI when payload is not valid
