+++ director/api/__init__.py
@@ -1,33 +1,16 @@
 from flask import jsonify, Blueprint
 from flask_json_schema import JsonValidationError
-from jsonschema.exceptions import ValidationError, SchemaError
-from jsonschema.validators import validator_for
+
+from director.utils import validate, format_schema_errors
 
 
 # Main application
 api_bp = Blueprint("api", __name__, url_prefix="/api")
 
 
-# Validate a payload according to a given schema
-def validate(payload, schema):
-    validator_cls = validator_for(schema)
-    validator = validator_cls(schema=schema)
-    errors = list(validator.iter_errors(payload))
-    if errors:
-        raise JsonValidationError("Payload is not valid", errors)
-
-
 @api_bp.errorhandler(JsonValidationError)
 def schema_exception_handler(e):
-    return (
-        jsonify(
-            {
-                "error": e.message,
-                "errors": [validation_err.message for validation_err in e.errors],
-            }
-        ),
-        400,
-    )
+    return jsonify(format_schema_errors(e)), 400
 
 
 @api_bp.errorhandler(404)
