+++ director/utils.py
@@ -1,0 +1,19 @@
+from flask_json_schema import JsonValidationError
+from jsonschema.validators import validator_for
+
+
+def validate(payload, schema):
+    """Validate a payload according to a given schema"""
+    validator_cls = validator_for(schema)
+    validator = validator_cls(schema=schema)
+    errors = list(validator.iter_errors(payload))
+    if errors:
+        raise JsonValidationError("Payload is not valid", errors)
+
+
+def format_schema_errors(e):
+    """Format FlaskJsonSchema validation errors"""
+    return {
+        "error": e.message,
+        "errors": [validation_err.message for validation_err in e.errors],
+    }
